move plugin selection to code generator
